+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Table row: selected states color update + documentation update
+
 ## [3.5.4][] - 2023-10-17
 
 ### Changed